graphics refactoring, slight graphics improvements
